Deleted LoginController.java. It was not used
